third_party/WebKit/LayoutTests/platform/win/http/tests/security/dangling-markup/option-expected.txt

CONSOLE ERROR: line 27: Form submission failed, as the <SELECT> element named 'dangling' was implicitly closed by reaching the end of the file. Please add an explicit end tag ('</SELECT>')
CONSOLE ERROR: line 27: Form submission failed, as the <SELECT> element named 'dangling' was implicitly closed by reaching the end of the file. Please add an explicit end tag ('</SELECT>')
CONSOLE ERROR: line 27: Form submission failed, as the <SELECT> element named 'dangling' was implicitly closed by reaching the end of the file. Please add an explicit end tag ('</SELECT>')
CONSOLE ERROR: line 27: Form submission failed, as the <SELECT> element named 'dangling' was implicitly closed by reaching the end of the file. Please add an explicit end tag ('</SELECT>')
CONSOLE ERROR: line 27: Form submission failed, as the <SELECT> element named 'dangling' was implicitly closed by reaching the end of the file. Please add an explicit end tag ('</SELECT>')
CONSOLE ERROR: line 27: Form submission failed, as the <SELECT> element named 'dangling' was implicitly closed by reaching the end of the file. Please add an explicit end tag ('</SELECT>')
This is a testharness.js-based test.
PASS       <form action="/security/resources/postmessage-post.php" method="post">        <input type="submit">        <select name="dangling"><option>     
PASS       <div>        <form action="/security/resources/postmessage-post.php" method="post">          <input type="submit">          <select name="dangling"><option>     
PASS       <form action="/security/resources/postmessage-post.php" method="post" id="form">        <input type="submit">      </form>      <select name="dangling" form="form"><option>     
PASS       <form action="/security/resources/postmessage-post.php" method="post">        <input type="submit">        <select name="dangling"><option label="yay">     
PASS       <div>        <form action="/security/resources/postmessage-post.php" method="post">          <input type="submit">          <select name="dangling"><option label="yay">     
PASS       <form action="/security/resources/postmessage-post.php" method="post" id="form">        <input type="submit">      </form>      <select name="dangling" form="form"><option label="yay">     
Harness: the test ran to completion.
